requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pytesseract
Pillow
dateparser
tzdata
